* commit '51a3b4c602579ff300f217dfd7859eb8d90adf62':
  SCA-377 - add testFetchReportsForPm
  SCA-377 - fix different types comparison

+++ modules/api/view-models/ReportsFetch.php
@@ -30,7 +30,7 @@ class ReportsFetch extends ViewModelAbstract
 {
     public function define()
     {
-        $usersId     = Yii::$app->request->getQueryParam('user_id');
+        $usersId     = (int) Yii::$app->request->getQueryParam('user_id');
         $date_period = Yii::$app->request->getQueryParam('date_period');
         $projectId   = Yii::$app->request->getQueryParam('project_id');
         $dateStart   = Yii::$app->request->getQueryParam('from_date');
